// src/_workspace/types/sct/SctCreateFromHistoryI.ts
export interface SctCreateFromHistoryI {
  SCT_CREATE_FROM_HISTORY_ID: number
  SCT_ID: string
  SCT_CREATE_FROM_SETTING_ID: number
  CREATE_FROM_SCT_ID: string
  CREATE_FROM_SCT_FISCAL_YEAR: number
  CREATE_FROM_SCT_PATTERN_ID: number
  CREATE_FROM_SCT_STATUS_PROGRESS_ID: number
  DESCRIPTION: string
  CREATE_BY: string
  CREATE_DATE: string
  UPDATE_BY: string
  UPDATE_DATE: string
  INUSE: string
}